requirement.txt
streamlit
facenet-pytorch
torch
opencv-python
Pillow
matplotlib
numpy
